+++ src/components/RollingPager/Card/PlusCard.jsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
-import colors from '../../styles/colors';
+import colors from '../../../styles/colors';
 import BaseCard from './BaseCard';
-import Circle from './Circle';
-import plusIcon from '../../assets/images/ic_plus.svg';
+import Circle from '../Circle';
+import plusIcon from '../../../assets/images/ic_plus.svg'
 
 const Card = styled(BaseCard)`
   display: flex;
